Work around prevail encoding size into fd

+++ src/ebpf/libs/api/test/droppacket.c
@@ -16,7 +16,7 @@ struct bpf_map_def port_map = {
       .size        = sizeof(struct bpf_map_def),
       .type        = 2,
       .key_size    = sizeof(__u32),
-      .value_size  = sizeof(__u32),
+      .value_size  = sizeof(__u64),
       .max_entries = 1
 };
